renamed vist-form.component to workspace
